Implement initial version of abstract file finder
I wanted to use the filter_directories() to not go into directories like
"build", "__pycache__", but it turned out to be very slow.

Other problem is that, if the name doesn't match, then all of the
headers are traversed, and it takes enormous amounts of time. The
workaround will be to not compile the header file using ClangTool, but
firstly check whether the file contains "class ClassName" or "struct
ClassName" string.