Reimplement extractor to match the abstract class
The matching for the abstract class is the same like in the abstract
class finder. In the MatchCallback we iterate over all bases (even
indirect) to seek for pure virtual functions. I wanted to create
matchers for methods:
    -> is pure and parent is the actual abstract class
    -> is pure and parent is base of the actual abstract class
but it is hard to create a matcher: "isBaseOf", because at the moment of
traversing the bases, we might not know yet whether it is a base of our
seeked abstract class.
The test pass at this stage, but the implementation is not yet fully
tested.