Develop suitable_place_in_class_finder a bit
Features added in this commit:
* When no public method is found, then a "public:" section is tried to
be found using ripgrep. In case no "public:" section is found within
the body, then ...
* ... we try to find the line, with the opening bracket of the class,
and we return it as the suitable place. This handles empty classes,
classes with private properties only, empty structs.

The function body line-span is taken from the source range of the class.

The LineFinder will skip any entities that have no definition (will skip
forward declarations of a class).

We have to inject the header file path to use the ripgrep tool properly.