Add initial Cucumber tests of implement interface code action
This will create a function which creates a cucumber test setup, where a
fixture runs the wire server with the cucumber tests. The wire server is
driver by the cucumber executable. The step definitions for the
implement interface code action is complete, but empty at this stage.
It will get filled soon.