Not use std::filesystem::create_directories()
The create_directories() ignores errors, and creates all the directories
on the way. Since keeping track of those directories, which were not
created, but already in place, is hard, then we prefer to use
create_directory().
In case multiple nested directories shall be created and RAII-deleted, then
use multiple DirectoryTree instances.