Retrieve full return type from template specialization
For template specialization as return type, we have explicitly request
fully qualified names for each of the arguments. Unfortunately, I
couldn't force libclang to do it for me automatically.

One more test case has been added, to verify whether it works.
Currently, only the return type is printed, but from the test output I
can see it is OK (because the test fails).

To properly compile std::variant, the "-std=gnu++20" flag had to be
added.