Add test with mutliple headers involved (FAILS)
This adds also a fixture, which neatly creates all the clang related
environment, to run the expander on.

Unfortunately, it fails, because the nested type is not properly
resolved: we get "External::Nested", and should obtain
"Namespace::External::Nested". We have to smartly desugar the nested
type, without making the library types, like "std::string" desugarized
to "std::basic_string<...>".

This bug applies both to return types and parameters.