Initial implement code action "implement interface"
This connects all the building blocks, the utilities developed recently,
which are the parts of this feature, to actually implement the
"implement interface code action". The filled tests pass so far.

I have discovered, that the SourceManager instance is different for each
of the Clang AST unit. I somewhat though that it is the same singleton
instance. It is not, so I have to refactor a little bit, to pass the
source manager tied to the node.

I have also taken the current indentation level for the implementor
class, with the help of the Lexer module.