Fix implementor mismatching due to wrong file choice
The implementor matcher didn't take into account that the header may
include other headers, and the source manager will create separate File
entries for them, thus the matches for the class with definition will
not match only the classes within the implementor file, but also the
included ones.
By adding a file path matcher, we fix this issue.