Fix base class matching
When multiple ASTs are used, and we want to check whether a base class
is already within the the base-clause, then matching by pointer will not
work.
We match by the fully qualified name, which might not be sufficient, for
some edge cases, e.g. when multiple different interfaces have the same
fully qualified name. On the other hand, this is a rare case, thus the
solution with the fully qualified name might be sufficient.