Refactor: hide implement interface code action under PIMPL
Because of that the interface node and the implementor node were
obtained early in the function and then passed around, then it means
that there was a good fit to introduce a class. But the main reason is
that the ast_units member was used implicitly within the member
functions, to get the SourceManager instances. Since this commit,
find_implementor() and find_interface() return a SourceManager instances
tied to the specific nodes.