Implement GenerateFunctionDefinition Code Action
This makes the same thing as the function_definition_generator, except
that it uses the full_function_declaration_expander, which handles more
cases (works more reliably and better, simply speaking), and it is able
to generate multiple function definitions, from a selected range.
The function_definition_generator will be using this code action as
well.