Fix scope removal from parameters when generating function
A new option has been introduced, that allows to remove scope only from the
parameters.
All tests pass with this commit.
I have fixed some compilation errors when invoking the clang tool (not
the compilation errors of this project). They appeared because the
"Given" steps within the Gherkin scenarios had missing semicolons, etc.
This is not a good fix, because actually the diagnostics shall not be
printed to stdout.
Added few more test cases which check whether the parameters are
properly shortified (the scope is removed from them).