Explicitly link against LLVM
When linking with libclang-cpp.so, we also need to link against
libllvm.so, because this library appears in the ELF header with a
DT_NEEDED tag.

Note that this problem does not appear when compiling with the
"-Wl,--copy-dt-needed-entries" option [0].

Otherwise an error like the following may appear:
/usr/bin/ld: src/suitable_place_in_class_finder/CMakeFiles/tsepepe_suitable_place_in_class_finder.dir/finder.cpp.o: undefined reference to symbol '_ZN4llvm3omp27isAllowedClauseForDirectiveENS0_9DirectiveENS0_6ClauseEj@@LLVM_15'
/usr/bin/ld: /usr/lib/libLLVM-15.so: error adding symbols: DSO missing from command line

Thus we add llvm explicitly to the linker commandline.

[0] https://stackoverflow.com/a/55086637