Fix clangTooling library redeclaration on Ubuntu
With Ubuntu 20.04 the 'clangTooling' library will be created by the call
to 'find_package(clang REQUIRED)'. Therefore, the call to
'add_library(clangTooling ...)' results in an error: ~"the target is
already defined". The test if the `clangTooling` TARGET already exists,
solves that on Ubuntu 20.04.
Test both for pair of packages: LLVM-14, Clang-14 and LLVM-15, Clang15.